Deal with most comments from review

+++ contracts/IReputationMiningCycle.sol
@@ -197,17 +197,17 @@ contract IReputationMiningCycle {
   function rewardStakersWithReputation(address[] stakers, address commonColonyAddress, uint reward, uint miningSkillId) public;
 
   /// @notice Get the timestamp that the current reputation mining window opened
-  function getReputationMiningWindowOpenTimestamp() public view returns (uint timestamp);
+  function getReputationMiningWindowOpenTimestamp() public view returns (uint256 timestamp);
 
   /// @notice Initialise this reputation mining cycle.
   /// @dev This will only be called once, by ColonyNetwork, in the same transaction that deploys this contract
   function initialise(address tokenLocking, address clnyToken) public;
 
   /// @notice Get the number of hashes that have been submitted this mining cycle
-  function getNSubmittedHashes() public view returns (uint nSubmittedHashes);
+  function getNSubmittedHashes() public view returns (uint256 nSubmittedHashes);
 
   /// @notice Get the number of hashes that have been invalidated this mining cycle
-  function getNInvalidatedHashes() public view returns (uint nInvalidatedHashes);
+  function getNInvalidatedHashes() public view returns (uint256 nInvalidatedHashes);
 
   /// @notice Get the address that made a particular submission
   /// @param hash The hash that was submitted
